# NeuroDesk/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The NeuroDesk Rest API.",
        "title": "NeuroDesk API",
        "contact": {
            "name": "NeuroDesk",
            "url": "https://localai.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/mudler/NeuroDesk/master/LICENSE"
        },
        "version": "2.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/router/decide": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "router"
                ],
                "summary": "Classify a prompt against a router model's policies (decision oracle)",
                "parameters": [
                    {
                        "description": "decide params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.RouterDecideRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.RouterDecideResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/router/{name}/corpus": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "router"
                ],
                "summary": "Seed the KNN routing corpus with labelled example prompts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "router model name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "labelled exemplars",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.RouterCorpusAddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.RouterCorpusAddResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "router"
                ],
                "summary": "Clear a router's KNN corpus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "router model name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.RouterCorpusClearResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/router/{name}/corpus/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "router"
                ],
                "summary": "Inspect a router's KNN corpus (label counts only, never texts)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "router model name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.RouterCorpusStatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.RouterCorpusAddRequest": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.RouterCorpusEntry"
                    }
                }
            }
        },
        "schema.RouterCorpusAddResponse": {
            "type": "object",
            "properties": {
                "added": {
                    "description": "Added is how many entries were embedded, persisted, and indexed.",
                    "type": "integer"
                },
                "label_counts": {
                    "description": "LabelCounts is the per-label exemplar count after the call.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "router": {
                    "type": "string"
                },
                "skipped": {
                    "description": "Skipped counts entries whose text was already in the corpus —\nduplicates are rejected rather than double-weighted.",
                    "type": "integer"
                },
                "total": {
                    "description": "Total is the corpus size after the call.",
                    "type": "integer"
                }
            }
        },
        "schema.RouterCorpusClearResponse": {
            "type": "object",
            "properties": {
                "cleared": {
                    "type": "integer"
                },
                "router": {
                    "type": "string"
                }
            }
        },
        "schema.RouterCorpusEntry": {
            "type": "object",
            "properties": {
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "schema.RouterCorpusStatsResponse": {
            "type": "object",
            "properties": {
                "embedding_model": {
                    "type": "string"
                },
                "embedding_models": {
                    "description": "EmbeddingModels lists the embedder fingerprints present in the\npersisted corpus; more than one means part of the corpus is\npending re-embedding on the next load.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "label_counts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "router": {
                    "type": "string"
                },
                "store_name": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "schema.RouterDecideNeighbor": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "similarity": {
                    "type": "number"
                }
            }
        },
        "schema.RouterDecideRequest": {
            "type": "object",
            "properties": {
                "input": {
                    "description": "Input is the user-visible prompt text to classify. Required.\nSchema-shape extraction (chat-message concatenation, etc.) is\nthe caller's responsibility — matches the Probe contract used\nby the in-band middleware.",
                    "type": "string"
                },
                "router": {
                    "description": "Router is the name of the router model (a ModelConfig with a\n`router:` block). Required.",
                    "type": "string"
                }
            }
        },
        "schema.RouterDecideResponse": {
            "type": "object",
            "properties": {
                "cache_similarity": {
                    "description": "CacheSimilarity carries the cosine similarity of the cache hit\n(0 when not cached).",
                    "type": "number"
                },
                "cached": {
                    "description": "Cached is true when the decision came from the L2 embedding\ncache rather than a fresh classifier run.",
                    "type": "boolean"
                },
                "candidate": {
                    "description": "Candidate is the model that would be routed to. Empty when no\ncandidate covers Labels AND no fallback is configured.",
                    "type": "string"
                },
                "classifier": {
                    "description": "Classifier is the classifier name that produced the decision\n(e.g. \"score\").",
                    "type": "string"
                },
                "fallback": {
                    "description": "Fallback is true when Candidate is the router's configured\nfallback because no candidate covered Labels. Lets callers\ndistinguish \"matched\" from \"fell back\" without comparing names.",
                    "type": "boolean"
                },
                "labels": {
                    "description": "Labels is the set of active policy labels.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "latency_ms": {
                    "description": "LatencyMs is the classifier's wall-clock cost.",
                    "type": "integer"
                },
                "nearest_similarity": {
                    "description": "NearestSimilarity is the cosine similarity of the closest KNN\ncorpus entry — populated by the knn classifier even when the\ndecision fell back because the probe was out of corpus range.\n0 for other classifiers.",
                    "type": "number"
                },
                "neighbors": {
                    "description": "Neighbors lists the corpus entries the knn classifier retrieved,\nby descending similarity, including ones below the similarity\ngate. Empty for other classifiers.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.RouterDecideNeighbor"
                    }
                },
                "router": {
                    "description": "Router echoes the requested router model.",
                    "type": "string"
                },
                "score": {
                    "description": "Score is the top label's softmax probability (the\nclassifier-side confidence signal).",
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}